# Improve TakeRange performance