```
## TakeRangeBench

| Method                                   | N       | Mean          | Error         | StdDev        | Gen0     | Gen1     | Gen2     | Allocated |
|----------------------------------------- |-------- |--------------:|--------------:|--------------:|---------:|---------:|---------:|----------:|
| FromArray_Linq                           | 1000000 |  5,521.188 us |  6,292.783 us |   344.9287 us |        - |        - |        - |     136 B |
| FromArray_ZLinq                          | 1000000 |  2,456.492 us |  9,128.028 us |   500.3380 us |        - |        - |        - |         - |
| FromArray_ZLinqPR                        | 1000000 |  1,899.951 us |  3,019.385 us |   165.5027 us |        - |        - |        - |         - |
|
| FromEnumerable_WithTake_Linq             | 1000000 |      9.353 us |     11.411 us |     0.6255 us |   0.0839 |        - |        - |    1424 B |
| FromEnumerable_WithTake_ZLinq            | 1000000 |  4,803.018 us |  6,782.172 us |   371.7537 us |        - |        - |        - |     136 B |
| FromEnumerable_WithTake_ZLinqPR          | 1000000 |      8.519 us |     23.307 us |     1.2775 us |        - |        - |        - |     136 B |

| FromEnumerable_Linq                      | 1000000 |  6,508.903 us |  5,749.797 us |   315.1658 us |        - |        - |        - |    1368 B |
| FromEnumerable_ZLinq                     | 1000000 | 10,966.269 us | 18,744.218 us | 1,027.4339 us |        - |        - |        - |     136 B |
| FromEnumerable_ZLinqPR                   | 1000000 |  8,076.553 us | 22,905.071 us | 1,255.5042 us |        - |        - |        - |     136 B |

| FromEnumerable_FromEnd_FromStart_Linq    | 1000000 | 10,374.365 us | 16,820.561 us |   921.9917 us | 500.0000 | 500.0000 | 500.0000 | 8389423 B |
| FromEnumerable_FromEnd_FromStart_ZLinq   | 1000000 | 10,268.761 us | 18,015.472 us |   987.4888 us |        - |        - |        - |     136 B |
| FromEnumerable_FromEnd_FromStart_ZLinqPR | 1000000 |  7,983.718 us |  2,019.655 us |   110.7041 us |        - |        - |        - |     136 B |

```
`Take(Range).Take` case was very slow.
Common type-independent logic should be implemented outside of generic types.