@neuecc 2025-07-08
I was also concerned about this area as it wasn't implemented well, so thank you for this excellent improvement!